+++ tests/test_convert.py
@@ -29,12 +29,17 @@ def test_converter(out_file, converter):
 
 
 @mark.parametrize('args', [
-    ['be_wa', 'tests/data-files/convert/be_wa/SIGEC_PARC_AGRI_ANON__2022_GEOPACKAGE_31370.zip|SIGEC_PARC_AGRI_ANON__2022.gpkg']
+    ['be_wa', ['tests/data-files/convert/be_wa/SIGEC_PARC_AGRI_ANON__2022_GEOPACKAGE_31370.zip|SIGEC_PARC_AGRI_ANON__2022.gpkg']],
+    ['ai4sf', ['tests/data-files/convert/ai4sf/1_vietnam_areas.gpkg', 'tests/data-files/convert/ai4sf/4_cambodia_areas.gpkg']]
 ])
 def test_converter_with_input(out_file, args):
-    converter, input_file = args
+    converter, input_files = args
     runner = CliRunner()
-    result = runner.invoke(convert, [converter, '-o', out_file.name, '-i', input_file])
+    args = [converter, '-o', out_file.name]
+    for input_file in input_files:
+        args.append('-i')
+        args.append(input_file)
+    result = runner.invoke(convert, args)
     assert result.exit_code == 0, result.output
     error = re.search('Skipped - |No schema defined', result.output)
     if error: